CmmSwitch: T783 regresses a lot
but it is the only one exhibiting that. The cause is the changed order
of branches in an if-then-else tree, which makes the hoople data flow
analysis traverse the blocks in a suboptimal order. Reverting that gets
rid of this regression, but has a consistent, if only very small
(+0.2%), negative effect on runtime. So I conclude that this test is an
extreme outlier and no reason to change the code.